Ignore ?complete when ?sha set
